requirements-base.txt
# ============================================
# OpenNotebookLM - 核心依赖（精简版）
# ============================================

# ------ Core Framework ------
fastapi==0.135.1
uvicorn==0.29.0
pydantic>=2.11.0
pydantic-settings>=2.12.0
python-dotenv==1.0.1
python-multipart>=0.0.22
httpx[socks]==0.28.1
requests==2.32.5
supabase==2.27.2

# ------ LLM / LangChain ------
openai==2.16.0
langchain==0.3.27
langchain-chroma==0.2.6
langchain-community==0.3.29
langchain-text-splitters==0.3.11

# ------ Vector Store / Embedding ------
chromadb==1.5.5
faiss-cpu==1.12.0
numpy<2.0.0

# ------ Document Parsing ------
PyMuPDF==1.26.7
pdfplumber==0.11.9
trafilatura>=2.0.0

# ------ Office / Export ------
python-pptx>=1.0.2
openpyxl>=3.1.5

# ------ Image ------
pillow>=10.4.0
matplotlib>=3.10.0

# ------ Database ------
sqlalchemy==2.0.25
sqlmodel==0.0.37
duckdb==1.5.0
pandas>=2.3.0

# ------ Web Scraping ------
beautifulsoup4>=4.14.0

# ------ Workflow Engine ------
pyyaml>=6.0.0
jinja2>=3.1.0
loguru>=0.7.3

# ------ Dev / Test ------
pytest>=9.0.0
